Reworked our User Database for common phrases.  It's now "officially obsolete".  They have now been moved to the Persistent Settings to better maintain and allow a path to stealth operation.  This also solves the problem of changing formats. Kept reading of existing kjvuser.s3db file (both user data storage and the template) to create an initial list.  Once our list exists in the CPersistentSettings, they will automatically override the older database storage, since PersistentSettings gets read after our application databases.  New phrases are only stored in PersistentSettings, as we are no longer writing the kjvuser.s3db file. Fixed updating of the CommonPhrasesCompleter list for all Search Phrases, which wasn't previously updating their models when another Search Phrase changed the list.  This then broke the add/delete button enables, which got fixed by their moving to PersistentSettings with a proper change notification signal. Reworked the ReadDB and BuildDB code to add/remove the database in the actual read/build functions rather than in the constructor/destructor for the database classes.  This also gets rid of the global QSqlDatabase objects -- and I hope it will get rid of the database memory cache issue we've continually had. Moved the User Phrases List variable to CPersistentSettings singleton instance object so that it's no longer a "global variable", except for the OSIS builder, which doesn't use CPersistentSettings (as an object).
git-svn-id: file:///home/dewhisna/Documents/programs/Bible/svndump/purebiblesearch_150731/code_gitcvt/trunk@729 523e89ec-37fb-4b40-b97f-42a194c95123

+++ programs/KJVCanOpener/BuildDB.cpp
@@ -46,9 +46,6 @@
 #include <QCoreApplication>
 #include <QFileInfo>
 
-QSqlDatabase g_sqldbBuildMain;
-QSqlDatabase g_sqldbBuildUser;
-
 #ifndef OSIS_PARSER_BUILD
 #include "main.h"
 extern CMyApplication *g_pMyApplication;
@@ -148,26 +145,13 @@ namespace {
 CBuildDatabase::CBuildDatabase(QWidget *pParent)
 	:	m_pParent(pParent)
 {
-	if (!g_sqldbBuildMain.contains(g_constrMainBuildConnection)) {
-		g_sqldbBuildMain = QSqlDatabase::addDatabase(g_constrDatabaseType, g_constrMainBuildConnection);
-	}
 
-	if (!g_sqldbBuildUser.contains(g_constrUserBuildConnection)) {
-		g_sqldbBuildUser = QSqlDatabase::addDatabase(g_constrDatabaseType, g_constrUserBuildConnection);
-	}
 }
 
 CBuildDatabase::~CBuildDatabase()
 {
-	if (g_sqldbBuildMain.contains(g_constrMainBuildConnection)) {
-		g_sqldbBuildMain = QSqlDatabase();
-		QSqlDatabase::removeDatabase(g_constrMainBuildConnection);
-	}
-
-	if (g_sqldbBuildUser.contains(g_constrUserBuildConnection)) {
-		g_sqldbBuildUser = QSqlDatabase();
-		QSqlDatabase::removeDatabase(g_constrUserBuildConnection);
-	}
+	if (QSqlDatabase::contains(g_constrMainBuildConnection)) QSqlDatabase::removeDatabase(g_constrMainBuildConnection);
+	if (QSqlDatabase::contains(g_constrUserBuildConnection)) QSqlDatabase::removeDatabase(g_constrUserBuildConnection);
 }
 
 // ============================================================================
@@ -911,6 +895,9 @@ bool CBuildDatabase::BuildPhrasesTable(bool bUserPhrases)
 			}
 
 			filePhrases.close();
+		} else {
+			// If this is the User Database, get it from the global list:
+			phrases.append(userPhrases());
 		}
 
 		// Create the table in the database:
@@ -923,10 +910,7 @@ bool CBuildDatabase::BuildPhrasesTable(bool bUserPhrases)
 			return false;
 		}
 
-		// Get the phrases and use the CPhraseListModel to sort it (since that will be displaying it later):
-		if (bUserPhrases) {
-			phrases.append(g_lstUserPhrases);
-		}
+		// Use the CPhraseListModel to sort it (since that will be displaying it later):
 		CPhraseListModel mdlPhrases(phrases);
 		mdlPhrases.sort(0, Qt::AscendingOrder);
 		phrases = mdlPhrases.phraseList();
@@ -961,29 +945,32 @@ bool CBuildDatabase::BuildPhrasesTable(bool bUserPhrases)
 
 bool CBuildDatabase::BuildDatabase(const QString &strDatabaseFilename)
 {
-	m_myDatabase = g_sqldbBuildMain;
+	m_myDatabase = QSqlDatabase::addDatabase(g_constrDatabaseType, g_constrMainBuildConnection);
 	m_myDatabase.setDatabaseName(strDatabaseFilename);
 
 //	QMessageBox::information(m_pParent, g_constrBuildDatabase, m_myDatabase.databaseName());
 
-	if (!m_myDatabase.open()) {
-		QMessageBox::warning(m_pParent, g_constrBuildDatabase, QObject::tr("Error: Couldn't open database file \"%1\".").arg(strDatabaseFilename));
-		return false;
-	}
-
 	bool bSuccess = true;
 
-	if ((!BuildDBInfoTable()) ||
-		(!BuildTestamentTable()) ||
-		(!BuildBooksTable()) ||
-		(!BuildChaptersTable()) ||
-		(!BuildVerseTables()) ||
-		(!BuildWordsTable()) ||
-		(!BuildFootnotesTables()) ||
-		(!BuildPhrasesTable(false))) bSuccess = false;
+	if (!m_myDatabase.open()) {
+		QMessageBox::warning(m_pParent, g_constrBuildDatabase, QObject::tr("Error: Couldn't open database file \"%1\".").arg(strDatabaseFilename));
+		bSuccess = false;
+	}
 
-	m_myDatabase.close();
+	if (bSuccess) {
+		if ((!BuildDBInfoTable()) ||
+			(!BuildTestamentTable()) ||
+			(!BuildBooksTable()) ||
+			(!BuildChaptersTable()) ||
+			(!BuildVerseTables()) ||
+			(!BuildWordsTable()) ||
+			(!BuildFootnotesTables()) ||
+			(!BuildPhrasesTable(false))) bSuccess = false;
+		m_myDatabase.close();
+	}
+
 	m_myDatabase = QSqlDatabase();
+	QSqlDatabase::removeDatabase(g_constrMainBuildConnection);
 
 	if (bSuccess) QMessageBox::information(m_pParent, g_constrBuildDatabase, QObject::tr("Build Complete!"));
 	return bSuccess;
@@ -991,25 +978,24 @@ bool CBuildDatabase::BuildDatabase(const QString &strDatabaseFilename)
 
 bool CBuildDatabase::BuildUserDatabase(const QString &strDatabaseFilename, bool bHideWarnings)
 {
-	m_myDatabase = g_sqldbBuildUser;
+	m_myDatabase = QSqlDatabase::addDatabase(g_constrDatabaseType, g_constrUserBuildConnection);
 	m_myDatabase.setDatabaseName(strDatabaseFilename);
+
+	bool bSuccess = true;
 
 	if (!m_myDatabase.open()) {
 		if (!bHideWarnings)
 			QMessageBox::warning(m_pParent, g_constrBuildDatabase, QObject::tr("Error: Couldn't open database file \"%1\".").arg(strDatabaseFilename));
-		return false;
-	}
-
-	bool bSuccess = true;
-
-	if (!BuildPhrasesTable(true)) {
 		bSuccess = false;
-	} else {
-		g_bUserPhrasesDirty = false;
 	}
 
-	m_myDatabase.close();
+	if (bSuccess) {
+		if (!BuildPhrasesTable(true)) bSuccess = false;
+		m_myDatabase.close();
+	}
+
 	m_myDatabase = QSqlDatabase();
+	QSqlDatabase::removeDatabase(g_constrUserBuildConnection);
 
 	return bSuccess;
 }
